+++ src/hotspot/cpu/arm/c1_CodeStubs_arm.cpp
@@ -53,22 +53,6 @@ void CounterOverflowStub::emit_code(LIR_Assembler* ce) {
   __ b(_continuation);
 }
 
-
-// TODO: ARM - is it possible to inline these stubs into the main code stream?
-
-
-RangeCheckStub::RangeCheckStub(CodeEmitInfo* info, LIR_Opr index, LIR_Opr array)
-  : _index(index), _array(array), _throw_index_out_of_bounds_exception(false) {
-  assert(info != NULL, "must have info");
-  _info = new CodeEmitInfo(info);
-}
-
-RangeCheckStub::RangeCheckStub(CodeEmitInfo* info, LIR_Opr index)
-  : _index(index), _array(), _throw_index_out_of_bounds_exception(true) {
-  assert(info != NULL, "must have info");
-  _info = new CodeEmitInfo(info);
-}
-
 void RangeCheckStub::emit_code(LIR_Assembler* ce) {
   __ bind(_entry);
 
@@ -191,16 +175,6 @@ void NewObjectArrayStub::emit_code(LIR_Assembler* ce) {
   ce->verify_oop_map(_info);
   __ b(_continuation);
 }
-
-
-// Implementation of MonitorAccessStubs
-
-MonitorEnterStub::MonitorEnterStub(LIR_Opr obj_reg, LIR_Opr lock_reg, CodeEmitInfo* info)
-: MonitorAccessStub(obj_reg, lock_reg)
-{
-  _info = new CodeEmitInfo(info);
-}
-
 
 void MonitorEnterStub::emit_code(LIR_Assembler* ce) {
   __ bind(_entry);
