+++ src/hotspot/cpu/riscv/c1_CodeStubs_riscv.cpp
@@ -68,18 +68,6 @@ void CounterOverflowStub::emit_code(LIR_Assembler* ce) {
   __ j(_continuation);
 }
 
-RangeCheckStub::RangeCheckStub(CodeEmitInfo* info, LIR_Opr index, LIR_Opr array)
-  : _index(index), _array(array), _throw_index_out_of_bounds_exception(false) {
-  assert(info != NULL, "must have info");
-  _info = new CodeEmitInfo(info);
-}
-
-RangeCheckStub::RangeCheckStub(CodeEmitInfo* info, LIR_Opr index)
-  : _index(index), _array(), _throw_index_out_of_bounds_exception(true) {
-  assert(info != NULL, "must have info");
-  _info = new CodeEmitInfo(info);
-}
-
 void RangeCheckStub::emit_code(LIR_Assembler* ce) {
   __ bind(_entry);
   if (_info->deoptimize_on_exception()) {
@@ -203,12 +191,6 @@ void NewObjectArrayStub::emit_code(LIR_Assembler* ce) {
   ce->verify_oop_map(_info);
   assert(_result->as_register() == x10, "result must in x10");
   __ j(_continuation);
-}
-
-// Implementation of MonitorAccessStubs
-MonitorEnterStub::MonitorEnterStub(LIR_Opr obj_reg, LIR_Opr lock_reg, CodeEmitInfo* info)
-: MonitorAccessStub(obj_reg, lock_reg) {
-  _info = new CodeEmitInfo(info);
 }
 
 void MonitorEnterStub::emit_code(LIR_Assembler* ce) {
