+++ src/hotspot/cpu/s390/c1_CodeStubs_s390.cpp
@@ -45,18 +45,6 @@ void C1SafepointPollStub::emit_code(LIR_Assembler* ce) {
   ShouldNotReachHere();
 }
 
-RangeCheckStub::RangeCheckStub(CodeEmitInfo* info, LIR_Opr index, LIR_Opr array)
-  : _index(index), _array(array), _throw_index_out_of_bounds_exception(false) {
-  assert(info != NULL, "must have info");
-  _info = new CodeEmitInfo(info);
-}
-
-RangeCheckStub::RangeCheckStub(CodeEmitInfo* info, LIR_Opr index)
-  : _index(index), _array(), _throw_index_out_of_bounds_exception(true) {
-  assert(info != NULL, "must have info");
-  _info = new CodeEmitInfo(info);
-}
-
 void RangeCheckStub::emit_code(LIR_Assembler* ce) {
   __ bind(_entry);
   if (_info->deoptimize_on_exception()) {
@@ -225,11 +213,6 @@ void NewObjectArrayStub::emit_code(LIR_Assembler* ce) {
   ce->verify_oop_map(_info);
   assert(_result->as_register() == Z_R2, "callee returns result in Z_R2,");
   __ z_brul(_continuation);
-}
-
-MonitorEnterStub::MonitorEnterStub(LIR_Opr obj_reg, LIR_Opr lock_reg, CodeEmitInfo* info)
-  : MonitorAccessStub(obj_reg, lock_reg) {
-  _info = new CodeEmitInfo(info);
 }
 
 void MonitorEnterStub::emit_code(LIR_Assembler* ce) {
